Struct dfareporting2::CampaignListMethodBuilder

pub struct CampaignListMethodBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

Retrieves a list of campaigns, possibly filtered.

A builder for the list method supported by a campaign resource. It is not used directly, but through a CampaignMethodsBuilder.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.campaigns().list("profileId")
             .subaccount_id("elitr")
             .sort_order("et")
             .sort_field("vero")
             .search_string("sadipscing")
             .page_token("sed")
             .overridden_event_tag_id("vero")
             .max_results(-91)
             .ids("magna")
             .excluded_ids("ea")
             .at_least_one_optimization_activity(false)
             .archived(true)
             .advertiser_ids("labore")
             .advertiser_group_ids("Lorem")
             .doit();

Methods

impl<'a, C, NC, A> CampaignListMethodBuilder<'a, C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>> + 'a, A: GetToken

fn doit(self) -> Result<CampaignsListResponse>

Perform the operation you have build so far.

fn profile_id(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the profile id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

User profile ID associated with this request.

fn subaccount_id(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the subaccount id query property to the given value.

Select only campaigns that belong to this subaccount.

fn sort_order(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the sort order query property to the given value.

Order of sorted results, default is ASCENDING.

fn sort_field(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the sort field query property to the given value.

Field by which to sort the list.

fn search_string(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the search string query property to the given value.

Allows searching for campaigns by name or ID. Wildcards (*) are allowed. For example, "campaign*2015" will return campaigns with names like "campaign June 2015", "campaign April 2015" or simply "campaign 2015". Most of the searches also add wildcards implicitly at the start and the end of the search string. For example, a search string of "campaign" will match campaigns with name "my campaign", "campaign 2015" or simply "campaign".

fn page_token(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the page token query property to the given value.

Value of the nextPageToken from the previous result page.

fn overridden_event_tag_id(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the overridden event tag id query property to the given value.

Select only campaigns that have overridden this event tag ID.

fn max_results(self, new_value: i32) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the max results query property to the given value.

Maximum number of results to return.

fn ids(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the ids query property to the given value.

Select only campaigns with these IDs.

fn excluded_ids(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the excluded ids query property to the given value.

Exclude campaigns with these IDs.

fn at_least_one_optimization_activity(self, new_value: bool) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the at least one optimization activity query property to the given value.

Select only campaigns that have at least one optimization activity.

fn archived(self, new_value: bool) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the archived query property to the given value.

Select only archived campaigns. Don't set this field to select both archived and non-archived campaigns.

fn advertiser_ids(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the advertiser ids query property to the given value.

Select only campaigns that belong to these advertisers.

fn advertiser_group_ids(self, new_value: &str) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the advertiser group ids query property to the given value.

Select only campaigns whose advertisers belong to these advertiser groups.

fn delegate(self, new_value: &'a mut Delegate) -> CampaignListMethodBuilder<'a, C, NC, A>

Sets the delegate property to the given value.

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

fn param<T>(self, name: T, value: T) -> CampaignListMethodBuilder<'a, C, NC, A> where T: Str

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> CampaignListMethodBuilder<'a, C, NC, A> where T: Str

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead of relying on the automated algorithm which simply prefers read-only scopes over those who are not.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.