Struct dfareporting2::CreativeAssignment

pub struct CreativeAssignment {
    pub weight: Option<i32>,
    pub sequence: Option<i32>,
    pub click_through_url: Option<ClickThroughUrl>,
    pub rich_media_exit_overrides: Vec<RichMediaExitOverride>,
    pub apply_event_tags: Option<bool>,
    pub creative_group_assignments: Vec<CreativeGroupAssignment>,
    pub companion_creative_overrides: Vec<CompanionClickThroughOverride>,
    pub start_time: Option<String>,
    pub ssl_compliant: Option<bool>,
    pub active: Option<bool>,
    pub creative_id: Option<String>,
    pub end_time: Option<String>,
    pub creative_id_dimension_value: Option<DimensionValue>,
}

Creative Assignment.

This type is not used in any activity, and only used as part of another schema.

Fields

weight

Weight of the creative assignment, applicable when the rotation type is CREATIVE_ROTATION_TYPE_RANDOM.

sequence

Sequence number of the creative assignment, applicable when the rotation type is CREATIVE_ROTATION_TYPE_SEQUENTIAL.

click_through_url

Click-through URL of the creative assignment.

rich_media_exit_overrides

Rich media exit overrides for this creative assignment. Applicable when the creative type is any of the following: - RICH_MEDIA_INPAGE - RICH_MEDIA_INPAGE_FLOATING - RICH_MEDIA_IM_EXPAND - RICH_MEDIA_EXPANDING - RICH_MEDIA_INTERSTITIAL_FLOAT - RICH_MEDIA_MOBILE_IN_APP - RICH_MEDIA_MULTI_FLOATING - RICH_MEDIA_PEEL_DOWN - ADVANCED_BANNER - VPAID_LINEAR - VPAID_NON_LINEAR

apply_event_tags

Whether applicable event tags should fire when this creative assignment is rendered. If this value is unset when the ad is inserted or updated, it will default to true for all creative types EXCEPT for INTERNAL_REDIRECT, INTERSTITIAL_INTERNAL_REDIRECT, and INSTREAM_VIDEO.

creative_group_assignments

Creative group assignments for this creative assignment. Only one assignment per creative group number is allowed for a maximum of two assignments.

companion_creative_overrides

Companion creative overrides for this creative assignment. Applicable to video ads.

start_time

Date and time that the assigned creative should start serving.

ssl_compliant

Whether the creative to be assigned is SSL-compliant. This is a read-only field that is auto-generated when the ad is inserted or updated.

active

Whether this creative assignment is active. When true, the creative will be included in the ad's rotation.

creative_id

ID of the creative to be assigned. This is a required field.

end_time

Date and time that the assigned creative should stop serving. Must be later than the start time.

creative_id_dimension_value

Dimension value for the ID of the creative. This is a read-only, auto-generated field.