Struct dfareporting2::CreativeGroup

pub struct CreativeGroup {
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub group_number: Option<i32>,
    pub advertiser_id: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub id: Option<String>,
    pub account_id: Option<String>,
}

Contains properties of a creative group.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

update (request|response)

patch (request|response)

list (none)

insert (request|response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#creativeGroup".

subaccount_id

Subaccount ID of this creative group. This is a read-only field that can be left blank.

name

Name of this creative group. This is a required field and must be less than 256 characters long and unique among creative groups of the same advertiser.

group_number

Subgroup of the creative group. Assign your creative groups to one of the following subgroups in order to filter or manage them more easily. This field is required on insertion and is read-only after insertion. Acceptable values are: - 1 - 2

advertiser_id

Advertiser ID of this creative group. This is a required field on insertion.

advertiser_id_dimension_value

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

id

ID of this creative group. This is a read-only, auto-generated field.

account_id

Account ID of this creative group. This is a read-only field that can be left blank.