Struct dfareporting2::Dfareporting

pub struct Dfareporting<C, NC, A> {
    // some fields omitted
}

Central instance to access all Dfareporting related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate dfareporting2;
use dfareporting2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2::Dfareporting;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.reports().list("profileId")
             .sort_order("sea")
             .sort_field("nonumy")
             .scope("dolores")
             .page_token("gubergren")
             .max_results(-95)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Methods

impl<'a, C, NC, A> Dfareporting<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn new(client: C, authenticator: A) -> Dfareporting<C, NC, A>

fn account_active_ad_summaries(&'a self) -> AccountActiveAdSummaryMethodsBuilder<'a, C, NC, A>

fn account_permission_groups(&'a self) -> AccountPermissionGroupMethodsBuilder<'a, C, NC, A>

fn account_permissions(&'a self) -> AccountPermissionMethodsBuilder<'a, C, NC, A>

fn account_user_profiles(&'a self) -> AccountUserProfileMethodsBuilder<'a, C, NC, A>

fn accounts(&'a self) -> AccountMethodsBuilder<'a, C, NC, A>

fn ads(&'a self) -> AdMethodsBuilder<'a, C, NC, A>

fn advertiser_groups(&'a self) -> AdvertiserGroupMethodsBuilder<'a, C, NC, A>

fn advertisers(&'a self) -> AdvertiserMethodsBuilder<'a, C, NC, A>

fn browsers(&'a self) -> BrowserMethodsBuilder<'a, C, NC, A>

fn campaign_creative_associations(&'a self) -> CampaignCreativeAssociationMethodsBuilder<'a, C, NC, A>

fn campaigns(&'a self) -> CampaignMethodsBuilder<'a, C, NC, A>

fn change_logs(&'a self) -> ChangeLogMethodsBuilder<'a, C, NC, A>

fn cities(&'a self) -> CityMethodsBuilder<'a, C, NC, A>

fn connection_types(&'a self) -> ConnectionTypeMethodsBuilder<'a, C, NC, A>

fn content_categories(&'a self) -> ContentCategoryMethodsBuilder<'a, C, NC, A>

fn countries(&'a self) -> CountryMethodsBuilder<'a, C, NC, A>

fn creative_assets(&'a self) -> CreativeAssetMethodsBuilder<'a, C, NC, A>

fn creative_field_values(&'a self) -> CreativeFieldValueMethodsBuilder<'a, C, NC, A>

fn creative_fields(&'a self) -> CreativeFieldMethodsBuilder<'a, C, NC, A>

fn creative_groups(&'a self) -> CreativeGroupMethodsBuilder<'a, C, NC, A>

fn creatives(&'a self) -> CreativeMethodsBuilder<'a, C, NC, A>

fn dimension_values(&'a self) -> DimensionValueMethodsBuilder<'a, C, NC, A>

fn directory_site_contacts(&'a self) -> DirectorySiteContactMethodsBuilder<'a, C, NC, A>

fn directory_sites(&'a self) -> DirectorySiteMethodsBuilder<'a, C, NC, A>

fn event_tags(&'a self) -> EventTagMethodsBuilder<'a, C, NC, A>

fn files(&'a self) -> FileMethodsBuilder<'a, C, NC, A>

fn floodlight_activities(&'a self) -> FloodlightActivityMethodsBuilder<'a, C, NC, A>

fn floodlight_activity_groups(&'a self) -> FloodlightActivityGroupMethodsBuilder<'a, C, NC, A>

fn floodlight_configurations(&'a self) -> FloodlightConfigurationMethodsBuilder<'a, C, NC, A>

fn landing_pages(&'a self) -> LandingPageMethodsBuilder<'a, C, NC, A>

fn metros(&'a self) -> MetroMethodsBuilder<'a, C, NC, A>

fn mobile_carriers(&'a self) -> MobileCarrierMethodsBuilder<'a, C, NC, A>

fn operating_system_versions(&'a self) -> OperatingSystemVersionMethodsBuilder<'a, C, NC, A>

fn operating_systems(&'a self) -> OperatingSystemMethodsBuilder<'a, C, NC, A>

fn placement_groups(&'a self) -> PlacementGroupMethodsBuilder<'a, C, NC, A>

fn placement_strategies(&'a self) -> PlacementStrategyMethodsBuilder<'a, C, NC, A>

fn placements(&'a self) -> PlacementMethodsBuilder<'a, C, NC, A>

fn platform_types(&'a self) -> PlatformTypeMethodsBuilder<'a, C, NC, A>

fn postal_codes(&'a self) -> PostalCodeMethodsBuilder<'a, C, NC, A>

fn regions(&'a self) -> RegionMethodsBuilder<'a, C, NC, A>

fn reports(&'a self) -> ReportMethodsBuilder<'a, C, NC, A>

fn sites(&'a self) -> SiteMethodsBuilder<'a, C, NC, A>

fn sizes(&'a self) -> SizeMethodsBuilder<'a, C, NC, A>

fn subaccounts(&'a self) -> SubaccountMethodsBuilder<'a, C, NC, A>

fn user_profiles(&'a self) -> UserProfileMethodsBuilder<'a, C, NC, A>

fn user_role_permission_groups(&'a self) -> UserRolePermissionGroupMethodsBuilder<'a, C, NC, A>

fn user_role_permissions(&'a self) -> UserRolePermissionMethodsBuilder<'a, C, NC, A>

fn user_roles(&'a self) -> UserRoleMethodsBuilder<'a, C, NC, A>