Struct dfareporting2::PathToConversionReportCompatibleFields

pub struct PathToConversionReportCompatibleFields {
    pub metrics: Vec<Metric>,
    pub per_interaction_dimensions: Vec<Dimension>,
    pub kind: Option<String>,
    pub conversion_dimensions: Vec<Dimension>,
    pub custom_floodlight_variables: Vec<Dimension>,
}

Represents fields that are compatible to be selected for a report of type "PATH_TO_CONVERSION".

This type is not used in any activity, and only used as part of another schema.

Fields

metrics

Metrics which are compatible to be selected in the "metricNames" section of the report.

per_interaction_dimensions

Per-interaction dimensions which are compatible to be selected in the "perInteractionDimensions" section of the report.

kind

The kind of resource this is, in this case dfareporting#pathToConversionReportCompatibleFields.

conversion_dimensions

Conversion dimensions which are compatible to be selected in the "conversionDimensions" section of the report.

custom_floodlight_variables

Custom floodlight variables which are compatible to be selected in the "customFloodlightVariables" section of the report.