Struct dfareporting2::SiteSettings

pub struct SiteSettings {
    pub disable_brand_safe_ads: Option<bool>,
    pub creative_settings: Option<CreativeSettings>,
    pub active_view_opt_out: Option<bool>,
    pub lookback_configuration: Option<LookbackConfiguration>,
    pub disable_new_cookie: Option<bool>,
    pub tag_setting: Option<TagSetting>,
}

Site Settings

This type is not used in any activity, and only used as part of another schema.

Fields

disable_brand_safe_ads

Whether brand safe ads are disabled for this site.

creative_settings

Site-wide creative settings.

active_view_opt_out

Whether active view creatives are disabled for this site.

lookback_configuration

Lookback window settings for this site.

disable_new_cookie

Whether new cookies are disabled for this site.

tag_setting

Configuration settings for dynamic and image floodlight tags.